# resources/docs/swagger.yaml
definitions:
...
  request.DictDataCreateRequest:
    properties:
      cssClass:
        description: 样式属性（其他样式扩展）
        type: string
      dictLabel:
        description: 字典标签
        type: string
      dictType:
        description: 字典类型编码
        type: string
      dictValue:
        description: 字典键值
        type: string
      listClass:
        description: 表格回显样式
        type: string
      orderNum:
        description: 字典排序
        minimum: 0
        type: integer
      remark:
        description: 备注
        type: string
    required:
    - dictLabel
    - dictType
    - dictValue
    type: object
  request.DictDataUpdateRequest:
    properties:
      cssClass:
        description: 样式属性（其他样式扩展）
        type: string
      dictLabel:
        description: 字典标签
        type: string
      dictValue:
        description: 字典键值
        type: string
      id:
        description: 字典编码
        type: integer
      listClass:
        description: 表格回显样式
        type: string
      orderNum:
        description: 字典排序
        minimum: 0
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - dictLabel
    - dictValue
    - id
    type: object
  request.DictTypeCreateRequest:
    properties:
      dictName:
        description: 字典类型名称
        type: string
      dictType:
        description: 字典类型编码
        type: string
      remark:
        description: 备注
        type: string
    required:
    - dictName
    - dictType
    type: object
  request.DictTypeUpdateRequest:
    properties:
      dictName:
        description: 字典类型名称
        type: string
      dictType:
        description: 字典类型编码
        type: string
      id:
        description: 字典主键
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - dictName
    - dictType
    - id
    type: object
...
  response.DictDataResponse:
    properties:
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      cssClass:
        description: 样式属性（其他样式扩展）
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      dictLabel:
        description: 字典标签
        type: string
      dictType:
        description: 字典类型编码
        type: string
      dictValue:
        description: 字典键值
        type: string
      id:
        description: 字典编码
        type: integer
      listClass:
        description: 表格回显样式
        type: string
      orderNum:
        description: 字典排序
        type: integer
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
  response.DictTypeResponse:
    properties:
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      dataList:
        description: 字典数据列表
        items:
          $ref: '#/definitions/response.DictDataResponse'
        type: array
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      dictName:
        description: 字典类型名称
        type: string
      dictType:
        description: 字典类型编码
        type: string
      id:
        description: 字典主键
        type: integer
      remark:
        description: 备注,
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
  response.FileResponse:
    properties:
      createBy:
        type: string
      createTime:
        type: string
      fileName:
        type: string
      filePath:
        type: string
      fileSize:
        type: integer
      fileUrl:
        type: string
      id:
        type: string
      mimeType:
        type: string
      remark:
        type: string
      userId:
        type: integer
    type: object
...
  response.RestRes:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: status ok
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample golang project
  title: Admin GOGOGO
  version: 0.0.1
paths:
...
  /dictData:
    post:
      consumes:
      - application/json
      description: 新增字典数据信息
      parameters:
      - description: 新增字典数据信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.DictDataCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典数据新增
      tags:
      - 系统字典数据
    put:
      consumes:
      - application/json
      description: 更新字典数据信息
      parameters:
      - description: 更新字典数据信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.DictDataUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典数据更新
      tags:
      - 系统字典数据
  /dictData/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除字典数据信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典数据伪删除
      tags:
      - 系统字典数据
    get:
      consumes:
      - application/json
      description: 详情字典数据信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DictDataResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典数据详情
      tags:
      - 系统字典数据
  /dictData/list:
    get:
      consumes:
      - application/json
      description: 获取所有字典数据信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DictDataResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典数据列表查询
      tags:
      - 系统字典数据
  /dictData/p:
    post:
      consumes:
      - application/json
      description: 获取所有字典数据信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DictDataResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典数据列表查询
      tags:
      - 系统字典数据
  /dictType:
    post:
      consumes:
      - application/json
      description: 新增字典类型信息
      parameters:
      - description: 新增字典类型信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.DictTypeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典类型新增
      tags:
      - 系统字典类型
    put:
      consumes:
      - application/json
      description: 更新字典类型信息
      parameters:
      - description: 更新字典类型信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.DictTypeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典类型更新
      tags:
      - 系统字典类型
  /dictType/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除字典类型信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典类型伪删除
      tags:
      - 系统字典类型
    get:
      consumes:
      - application/json
      description: 详情字典类型信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DictTypeResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典类型详情
      tags:
      - 系统字典类型
  /dictType/list:
    get:
      consumes:
      - application/json
      description: 获取所有字典类型信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DictTypeResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典类型列表查询
      tags:
      - 系统字典类型
  /dictType/p:
    post:
      consumes:
      - application/json
      description: 获取所有字典类型信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.DictTypeResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 字典类型列表查询
      tags:
      - 系统字典类型
  /file/:id:
    delete:
      consumes:
      - application/json
      description: 文件上传信息删除
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文件上传删除
      tags:
      - 文件上传
    get:
      consumes:
      - application/json
      description: 文件上传详情信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.FileResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文件上传详情
      tags:
      - 文件上传
  /file/p:
    post:
      consumes:
      - application/json
      description: 分页获取所有文件上传信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.FileResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文件上传分页查询
      tags:
      - 文件上传
  /file/upload/common:
    post:
      consumes:
      - application/json
      description: 文件上传信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.FileResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 文件上传
      tags:
      - 文件上传
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"